+++ poetry/console/styles/poetry.py
@@ -4,17 +4,11 @@ from cleo.styles import OutputStyle
 
 class PoetryStyle(CleoStyle):
 
-    def __init__(self, i, o, venv):
-        self._venv = venv
-
+    def __init__(self, i, o):
         super().__init__(i, o)
 
         self.output.get_formatter().add_style('warning', 'black', 'yellow')
         self.output.get_formatter().add_style('question', 'blue')
-
-    @property
-    def venv(self):
-        return self._venv
 
     def writeln(self, messages,
                 type=OutputStyle.OUTPUT_NORMAL,