+++ tests/masonry/builders/test_wheel.py
@@ -6,6 +6,7 @@ from pathlib import Path
 from poetry import Poetry
 from poetry.io import NullIO
 from poetry.masonry.builders import WheelBuilder
+from poetry.utils.venv import NullVenv
 
 
 fixtures_dir = Path(__file__).parent / 'fixtures'
@@ -28,7 +29,7 @@ def clear_samples_dist():
 
 def test_wheel_module():
     module_path = fixtures_dir / 'module1'
-    WheelBuilder.make(Poetry.create(str(module_path)), NullIO())
+    WheelBuilder.make(Poetry.create(str(module_path)), NullVenv(), NullIO())
 
     whl = module_path / 'dist' / 'module1-0.1-py2.py3-none-any.whl'
 
@@ -37,7 +38,7 @@ def test_wheel_module():
 
 def test_wheel_package():
     module_path = fixtures_dir / 'complete'
-    WheelBuilder.make(Poetry.create(str(module_path)), NullIO())
+    WheelBuilder.make(Poetry.create(str(module_path)), NullVenv(), NullIO())
 
     whl = module_path / 'dist' / 'my_package-1.2.3-py3-none-any.whl'
 
@@ -46,7 +47,7 @@ def test_wheel_package():
 
 def test_wheel_prerelease():
     module_path = fixtures_dir / 'prerelease'
-    WheelBuilder.make(Poetry.create(str(module_path)), NullIO())
+    WheelBuilder.make(Poetry.create(str(module_path)), NullVenv(), NullIO())
 
     whl = module_path / 'dist' / 'prerelease-0.1b1-py2.py3-none-any.whl'